moved table from containers to components folder; fixed key index warnings
