+++ client/src/components/table/cell/line-chart-cell/component.tsx
