+++ client/src/components/table/data-row/index.ts
