+++ client/src/components/table/enums.ts
