+++ client/src/components/table/group-row/component.tsx
