+++ client/src/components/table/summary-row/component.tsx
